ml/daytrading/requirements.txt
lightgbm>=4.0.0
pandas>=2.1.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
pyarrow>=14.0.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
tqdm>=4.66.0
